backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
flask-bcrypt==1.0.1
pymongo==4.7.3
python-dotenv==1.0.1
apscheduler==3.10.4
scikit-learn>=1.5.0
pandas>=2.2.2
numpy>=1.26.4
joblib>=1.4.2
reportlab==4.2.0
Werkzeug==3.0.3
